Change return data type for jumlahPerProduk method from Double to Integer

+++ src/main/java/com/ecommerce/web/jpa/e_commerce_web_jpa/Repositories/OmsetRespository.java
@@ -19,7 +19,7 @@ public interface OmsetRespository extends JpaRepository<Omset, Integer> {
     Double jumlahHargaPerProduk(String idProduk);
 
     @Query("SELECT o.jumlahPenjualan  FROM Omset o JOIN o.idProduk p WHERE p.id = :idProduk")
-    Double jumlahPerProdukTerjual(String idProduk);
+    Integer jumlahPerProdukTerjual(String idProduk);
 
     @Query("SELECT SUM(o.jumlahPenjualan * p.harga) FROM Omset o JOIN o.idProduk p")
     Double jumlahOmset();
